requirements.txt
# ── Core ──
streamlit>=1.38.0,<2.0
PyYAML>=6.0,<7.0
requests>=2.31,<3.0

# ── Document parsing ──
PyMuPDF>=1.24,<2.0
python-docx>=1.1,<2.0

# ── PageIndex (self-hosted) ──
pageindex>=0.2.8
litellm>=1.80,<2.0
python-dotenv>=1.0,<2.0

# ── Vector store ──
chromadb>=0.5,<1.0

# ── Export ──
fpdf2>=2.8,<3.0

# ── Testing ──
pytest>=8.0,<9.0
pytest-cov>=5.0,<6.0
